+++ index.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Awesome Book | Home</title>
     <!-- CSS Link -->
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="index.css">
     <!-- google font -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -26,12 +26,10 @@
                     <li id="addNew">Add new</li>
                     <li id="contact">Contact</li>
                 </ul>
+
             </div>
         </nav>
-        <div class="container">
-            <p id="time"></p>
-            <p id="message"></p>
-        </div>
+        <p id="time_tag"></p>
     </header>
     <!-- end of header -->
 
@@ -70,9 +68,9 @@
                     </ul>
                 </div>
             </section>
+            
     </main>
     <!-- end of main -->
-
     <!-- start of footer -->
     <footer>
         <div class="container">
@@ -80,10 +78,8 @@
         </div>
     </footer>
     <!-- end of footer -->
-
-    <!-- js link -->
-    <script src="js/main.js"></script>
-    <script src="js/date.js"></script>
+        <!-- js link -->
+        <script type="module" src="index.js"></script>
 </body>
 
 </html>
